Déplace réduction Pinel vers son propre fichier
